+++ src/main/java/com/ooa1769/bs/book/support/BookMarkRepository.java
@@ -1,0 +1,7 @@
+package com.ooa1769.bs.book.support;
+
+import com.ooa1769.bs.book.BookMark;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookMarkRepository extends JpaRepository<BookMark, Long> {
+}
