+++ src/main/java/com/ooa1769/bs/support/security/SecurityUserDetailService.java
@@ -1,7 +1,8 @@
 package com.ooa1769.bs.support.security;
 
-import com.ooa1769.bs.domain.member.Member;
-import com.ooa1769.bs.respository.member.MemberRepository;
+import com.ooa1769.bs.member.Member;
+import com.ooa1769.bs.member.support.MemberRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
@@ -19,8 +20,9 @@ public class SecurityUserDetailService implements UserDetailsService {
 
     private static final String ROLE_USER = "ROLE_USER";
 
-    private MemberRepository memberRepository;
+    private final MemberRepository memberRepository;
 
+    @Autowired
     public SecurityUserDetailService(MemberRepository memberRepository) {
         this.memberRepository = memberRepository;
     }
@@ -31,11 +33,10 @@ public class SecurityUserDetailService implements UserDetailsService {
         Member member = memberOpt.orElseThrow(
                 () -> new UsernameNotFoundException("회원이 존재하지 않습니다."));
 
-        return new User(member.getEmail(), member.getPassword(), member.isEnabled(), true,true,true, getAuthorities(ROLE_USER));
+        return new User(member.getEmail(), member.getPassword(), member.isEnabled(), true, true, true, getAuthorities(ROLE_USER));
     }
 
     private Collection<? extends GrantedAuthority> getAuthorities(String role) {
         return Arrays.asList(new SimpleGrantedAuthority(role));
     }
-
 }
