+++ src/main/java/com/ooa1769/bs/web/BookRestController.java
@@ -1,10 +1,11 @@
 package com.ooa1769.bs.web;
 
-import com.ooa1769.bs.domain.book.SearchOption;
-import com.ooa1769.bs.domain.member.Member;
-import com.ooa1769.bs.service.book.BookService;
-import com.ooa1769.bs.service.member.MemberService;
+import com.ooa1769.bs.book.SearchOption;
+import com.ooa1769.bs.book.support.BookService;
+import com.ooa1769.bs.member.Member;
+import com.ooa1769.bs.member.support.MemberService;
 import com.ooa1769.bs.support.security.LoginMember;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,9 +16,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/books")
 public class BookRestController {
 
-    private BookService bookService;
-    private MemberService memberService;
+    private final BookService bookService;
+    private final MemberService memberService;
 
+    @Autowired
     public BookRestController(BookService bookService, MemberService memberService) {
         this.bookService = bookService;
         this.memberService = memberService;
@@ -26,7 +28,7 @@ public class BookRestController {
     @GetMapping("/search")
     public ResponseEntity<?> search(@LoginMember Member member, SearchOption searchOption) {
         memberService.addSearchHistory(member, searchOption.getQuery());
-        return new ResponseEntity<>(bookService.search(searchOption), HttpStatus.OK);
+        return new ResponseEntity<>(bookService.getBooks(searchOption), HttpStatus.OK);
     }
 
     @GetMapping("/keywords")
