+++ src/main/java/com/ooa1769/bs/web/dto/MemberDto.java
@@ -1,9 +1,11 @@
 package com.ooa1769.bs.web.dto;
 
-import com.ooa1769.bs.domain.member.Member;
-import lombok.Data;
+import com.ooa1769.bs.member.Member;
+import lombok.Getter;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 public class MemberDto {
 
     private String email;
