+++ src/main/resources/application.properties
@@ -13,5 +13,6 @@ spring.jpa.database-platform=H2
 
 logging.level.org.hibernate.type=trace
 logging.level.org.apache.http=debug
+#logging.level.org.springframework.security.web=debug
 
 spring.thymeleaf.cache=false
