+++ src/test/java/com/ooa1769/bs/book/BookMarkTest.java
@@ -1,0 +1,11 @@
+package com.ooa1769.bs.book;
+
+import org.junit.Test;
+
+public class BookMarkTest {
+
+    @Test
+    public void 북마크_추가() throws Exception {
+
+    }
+}
