--- src/test/java/com/ooa1769/bs/domain/member/MemberTest.java
@@ -1,5 +1,0 @@
-package com.ooa1769.bs.domain.member;
-
-public class MemberTest {
-
-}
